+++ tests/oidc/test_errors.py
@@ -5,3 +5,57 @@ def test_json_oauth_errors(testclient):
         "error": "not_found",
         "error_description": "The requested URL was not found on the server. If you entered the URL manually please check your spelling and try again.",
     }
+
+
+def test_invalid_client(testclient, logged_user, keypair):
+    res = testclient.get(
+        "/oauth/authorize",
+        params=dict(
+            response_type="code",
+            client_id="invalid",
+            scope="openid profile email groups address phone",
+            nonce="somenonce",
+            redirect_uri="https://client.test/redirect1",
+        ),
+        status=400,
+    )
+    assert res.json == {
+        "error": "invalid_client",
+        "error_description": "Invalid client.",
+    }
+
+
+def test_consent_with_no_redirect_uri(testclient, logged_user, client, backend):
+    res = testclient.get(
+        "/oauth/authorize",
+        params=dict(
+            response_type="code",
+            client_id=client.client_id,
+            nonce="somenonce",
+        ),
+        status=400,
+    )
+    assert res.json == {
+        "error": "invalid_request",
+        "error_description": "Missing 'redirect_uri' in request.",
+    }
+
+
+def test_missing_client_id(
+    testclient, logged_user, client, keypair, trusted_client, backend
+):
+    """Missing client_id should raise a 400 error."""
+    res = testclient.get(
+        "/oauth/authorize",
+        params=dict(
+            response_type="code",
+            scope="openid profile email groups address phone",
+            nonce="somenonce",
+            redirect_uri="https://client.test/redirect1",
+        ),
+        status=400,
+    )
+    assert res.json == {
+        "error": "invalid_request",
+        "error_description": "'client_id' parameter is missing.",
+    }